renaming respository to service
